requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pycodestyle==2.10.0
pydantic==1.10.7
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
